add WorldActivity class, update fetchVisitor flow
