+++ clients/client-topia/src/controllers/__tests__/visitor.test.ts
@@ -1,10 +1,11 @@
 import MockAdapter from "axios-mock-adapter";
-import { visitors } from "../../__mocks__";
+import { visitor } from "__mocks__";
 import { Visitor as VisitorClass, Topia } from "controllers";
 import { VisitorFactory } from "factories";
 
 const apiDomain = "api.topia.io";
-const id = visitors["1"].playerId;
+const id = visitor.playerId;
+const urlSlug = "exampleWorld";
 
 describe("Visitor Class", () => {
   let mock: MockAdapter, testVisitor: VisitorClass, topia: Topia, Visitor: VisitorFactory;
@@ -16,7 +17,7 @@ describe("Visitor Class", () => {
     });
     mock = new MockAdapter(topia.axios);
     Visitor = new VisitorFactory(topia);
-    testVisitor = Visitor.create(id, "exampleWorld");
+    testVisitor = Visitor.create(id, urlSlug);
   });
 
   afterEach(() => {
@@ -24,8 +25,14 @@ describe("Visitor Class", () => {
     jest.resetAllMocks();
   });
 
+  it("should update visitor details", async () => {
+    mock.onGet(`https://${apiDomain}/api/world/exampleWorld/visitors/1`).reply(200, visitor);
+    await testVisitor.fetchVisitor();
+    expect(mock.history.get.length).toBe(1);
+  });
+
   it("should move a list of visitors to uniquely specified coordinates", async () => {
-    mock.onPut(`https://${apiDomain}/api/world/exampleWorld/visitors/${id}/move`).reply(200);
+    mock.onPut(`https://${apiDomain}/api/world/${urlSlug}/visitors/${id}/move`).reply(200);
     await testVisitor.moveVisitor({
       shouldTeleportVisitor: true,
       x: 100,
