+++ clients/client-topia/src/factories/UserFactory.ts
@@ -11,12 +11,6 @@ export class UserFactory {
   create(options?: UserOptionalInterface): User {
     return new User(this.topia, options);
   }
-
-  async get(options?: UserOptionalInterface): Promise<User> {
-    const user = new User(this.topia, options);
-    await user.fetchUserByVisitorId();
-    return user;
-  }
 }
 
 export default UserFactory;
