+++ clients/client-topia/src/factories/VisitorFactory.ts
@@ -8,16 +8,14 @@ export class VisitorFactory {
     this.topia = topia;
   }
 
-  async get(id: number, urlSlug: string, options?: VisitorOptionalInterface): Promise<Visitor> {
-    const visitor = new Visitor(this.topia, id, urlSlug, options);
-
-    await visitor.fetchVisitor();
-
-    return visitor;
-  }
-
   create(id: number, urlSlug: string, options?: VisitorOptionalInterface): Visitor {
     return new Visitor(this.topia, id, urlSlug, options);
+  }
+
+  async get(id: number, urlSlug: string, options?: VisitorOptionalInterface): Promise<Visitor> {
+    const visitor = new Visitor(this.topia, id, urlSlug, options);
+    await visitor.fetchVisitor();
+    return visitor;
   }
 }
