+++ clients/client-topia/src/factories/WorldActivityFactory.ts
@@ -1,0 +1,16 @@
+import { Topia, WorldActivity } from "controllers";
+import { WorldOptionalInterface } from "interfaces";
+
+export class WorldActivityFactory {
+  topia: Topia;
+
+  constructor(topia: Topia) {
+    this.topia = topia;
+  }
+
+  create(urlSlug: string, options?: WorldOptionalInterface): WorldActivity {
+    return new WorldActivity(this.topia, urlSlug, options);
+  }
+}
+
+export default WorldActivityFactory;
