+++ clients/client-topia/src/factories/index.ts
@@ -3,4 +3,5 @@ export { DroppedAssetFactory } from "./DroppedAssetFactory";
 export { SceneFactory } from "./SceneFactory";
 export { UserFactory } from "./UserFactory";
 export { VisitorFactory } from "./VisitorFactory";
+export { WorldActivityFactory } from "./WorldActivityFactory";
 export { WorldFactory } from "./WorldFactory";
