+++ clients/client-topia/src/index.ts
@@ -1,5 +1,12 @@
 export { Topia } from "controllers";
-export { AssetFactory, DroppedAssetFactory, UserFactory, VisitorFactory, WorldFactory } from "factories";
+export {
+  AssetFactory,
+  DroppedAssetFactory,
+  UserFactory,
+  VisitorFactory,
+  WorldActivityFactory,
+  WorldFactory,
+} from "factories";
 
 Error.stackTraceLimit = 20;
 process.on("unhandledRejection", (reason: any) => {
