+++ clients/client-topia/src/interfaces/WorldActivityInterfaces.ts
@@ -1,0 +1,13 @@
+import { InteractiveCredentials } from "types";
+
+export interface WorldActivityOptionalInterface {
+  credentials?: InteractiveCredentials | object;
+}
+
+export interface MoveAllVisitorsInterface {
+  shouldFetchVisitors?: boolean;
+  shouldTeleportVisitors?: boolean;
+  scatterVisitorsBy?: number;
+  x: number;
+  y: number;
+}
