+++ clients/client-topia/src/interfaces/WorldInterfaces.ts
@@ -32,11 +32,3 @@ export interface WorldOptionalInterface {
   attributes?: WorldInterface | object;
   credentials?: InteractiveCredentials | object;
 }
-
-export interface MoveAllVisitorsInterface {
-  shouldFetchVisitors?: boolean;
-  shouldTeleportVisitors?: boolean;
-  scatterVisitorsBy?: number;
-  x: number;
-  y: number;
-}
