+++ clients/client-topia/src/interfaces/index.ts
@@ -5,4 +5,5 @@ export * from "./SDKInterfaces";
 export * from "./TopiaInterfaces";
 export * from "./UserInterfaces";
 export * from "./VisitorInterfaces";
+export * from "./WorldActivityInterfaces";
 export * from "./WorldInterfaces";
